- adding disabled dates in db from employee - get disabled dates by employeeId - get booked days by employee id

+++ homeMaintenance.Domain/Entities/BookingInfo.cs
@@ -6,7 +6,7 @@ namespace homeMaintenance.Domain.Entities
     {
         public string FullName { get; set; }
         public string Avatar { get; set; }
-        public BigInteger Amount { get; set; }
+        public long Amount { get; set; }
         public DateTime StartDateTime { get; set; }
         public DateTime EndDateTime { get; set; }
